# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Job Portal server. You can visit the GitHub repository at https://github.com/fazil/Job_Portal_Project",
        "title": "Go + Gin Workey API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/worker/addprofile": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Add profile for Worker",
                "operationId": "worker add profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name : ",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gender : ",
                        "name": "gender",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date Of Birth : ",
                        "name": "dateofbirth",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "House Name : ",
                        "name": "housename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Place : ",
                        "name": "place",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post : ",
                        "name": "post",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pin : ",
                        "name": "pin",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact Number : ",
                        "name": "contactnumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email Id : ",
                        "name": "emailid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Photo : ",
                        "name": "photo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/worker/changepassword": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Change Password for worker",
                "operationId": "worker change password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name : ",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gender : ",
                        "name": "gender",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date Of Birth : ",
                        "name": "dateofbirth",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "House Name : ",
                        "name": "housename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Place : ",
                        "name": "place",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post : ",
                        "name": "post",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pin : ",
                        "name": "pin",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact Number : ",
                        "name": "contactnumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email Id : ",
                        "name": "emailid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Photo : ",
                        "name": "photo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/worker/editprofile": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Edit profile for Worker",
                "operationId": "worker edit profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name : ",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gender : ",
                        "name": "gender",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date Of Birth : ",
                        "name": "dateofbirth",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "House Name : ",
                        "name": "housename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Place : ",
                        "name": "place",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post : ",
                        "name": "post",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pin : ",
                        "name": "pin",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact Number : ",
                        "name": "contactnumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email Id : ",
                        "name": "emailid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Photo : ",
                        "name": "photo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}